struct: add schemas folder , move Login and Signup schema to schemas folder
